Move cloudinary implementation code into package
