+++ .changeset/eleven-owls-give.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/cloudinary': patch
+---
+
+Removed the dependency on the legacy `CloudinaryImage` field implementation.
