+++ .changeset/good-weeks-study.md
@@ -1,5 +1,5 @@
 ---
-'@keystone-next/fields-cloudinary-image-legacy': patch
+'@keystone-next/cloudinary': patch
 ---
 
 Refactored the core implementation to not depend on the `File` type.
