+++ package.json
@@ -121,8 +121,6 @@
       "packages/session",
       "packages/test-utils",
       "packages/utils",
-      "packages/fields-auto-increment",
-      "packages/fields-cloudinary-image",
       "packages/fields",
       "packages/adapter-prisma",
       "packages/server-side-graphql-client",
