+++ packages-next/cloudinary/src/test-files/keystone.jpeg
