--- packages/fields-cloudinary-image/.env.example
@@ -1,4 +1,0 @@
-# Setup the following in your .env file
-# CLOUDINARY_CLOUD_NAME=cloud-name
-# CLOUDINARY_KEY=cloud-key
-# CLOUDINARY_SECRET=cloud-secret
