--- packages/fields-cloudinary-image/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@keystone-next/fields-cloudinary-image-legacy",
-  "description": "KeystoneJS Cloudinary Image Field Type",
-  "version": "5.0.0",
-  "author": "The KeystoneJS Development Team",
-  "license": "MIT",
-  "repository": "https://github.com/keystonejs/keystone/tree/master/packages/fields-cloudinary-image",
-  "homepage": "https://github.com/keystonejs/keystone",
-  "engines": {
-    "node": ">=10.0.0"
-  },
-  "dependencies": {
-    "@keystone-next/adapter-prisma-legacy": "^5.0.0",
-    "@keystone-next/fields-legacy": "^25.0.0",
-    "cuid": "^2.1.8"
-  },
-  "main": "dist/fields-cloudinary-image-legacy.cjs.js",
-  "module": "dist/fields-cloudinary-image-legacy.esm.js",
-  "devDependencies": {
-    "@keystone-next/cloudinary": "*",
-    "@keystone-next/fields": "*",
-    "@keystone-next/file-adapters-legacy": "*",
-    "cloudinary": "^1.25.1",
-    "dotenv": "^8.2.0",
-    "graphql-upload": "^11.0.0",
-    "mime": "^2.5.2"
-  }
-}
