--- packages/fields-cloudinary-image/src/index.js
@@ -1,12 +1,0 @@
-import {
-  CloudinaryImage as Implementation,
-  PrismaCloudinaryImageInterface,
-} from './Implementation';
-
-export const CloudinaryImage = {
-  type: 'CloudinaryImage',
-  implementation: Implementation,
-  adapters: {
-    prisma: PrismaCloudinaryImageInterface,
-  },
-};
